Extract algorithm selection
Parts concerning the algorithm selection were delegated to the new fxml
and controller pair of "AlgorithmSelection". Needed adjustments in
Mediator.java and main.fxml were also made.
