+++ src/main/resources/com/targus/algorithmSelection.fxml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.collections.FXCollections?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import java.lang.*?>
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="com.targus.ui.controllers.AlgorithmSelectionController">
+
+    <ComboBox fx:id="algorithmComboBox" onAction="#handleAlgorithmSelection" value="Please select algorithm">
+        <items>
+            <FXCollections fx:factory="observableArrayList">
+                <String fx:value="Standard GA"/>
+                <String fx:value="Improved GA"/>
+                <String fx:value="Simulated Annealing"/>
+                <String fx:value="Greedy Algorithm"/>
+            </FXCollections>
+        </items>
+    </ComboBox>
+    <StackPane fx:id="algorithmInputs" style="-fx-background-color: #f2f2f2;" visible="false">
+        <VBox fx:id="GAInputs" spacing="20">
+            <HBox alignment="CENTER_LEFT">
+                <Label text="Mutation Operator: " prefWidth="128"/>
+                <ComboBox fx:id="mutationComboBox" value="OneBitMutation" prefWidth="128">
+                    <items>
+                        <FXCollections fx:factory="observableArrayList">
+                            <String fx:value="OneBitMutation"/>
+                            <String fx:value="KBitMutation"/>
+                        </FXCollections>
+                    </items>
+                </ComboBox>
+            </HBox>
+            <HBox>
+                <Label text="Mutation Rate (in %): " prefWidth="128"/>
+                <TextField fx:id="mutationRateTextField" prefWidth="128" alignment="CENTER"/>
+            </HBox>
+            <HBox alignment="CENTER_LEFT">
+                <Label text="Termination Type: " prefWidth="128"/>
+                <ComboBox fx:id="terminationComboBox" prefWidth="128" value="Time Based">
+                    <items>
+                        <FXCollections fx:factory="observableArrayList">
+                            <String fx:value="Time Based"/>
+                            <String fx:value="Iteration Based"/>
+                        </FXCollections>
+                    </items>
+                </ComboBox>
+            </HBox>
+            <HBox>
+                <Label text="Termination value:  " prefWidth="128"/>
+                <TextField fx:id="terminationTextField" prefWidth="128" alignment="CENTER"/>
+            </HBox>
+        </VBox>
+    </StackPane>
+</AnchorPane>
